Move modalCleanup from util to device.
