+++ src/components/ModalComponent.native.tsx
@@ -2,7 +2,7 @@
 import React from 'react'
 /* import classNames from 'classnames'
 import { FADEOUT_DURATION, MODAL_CLOSE_DURATION } from '../constants'
-import { modalCleanup } from '../util'
+import modalCleanup from '../device/modalCleanup'
 import { Connected } from '../@types' */
 import { /* closeModal */ modalComplete /* tutorial */ } from '../action-creators'
 import { useDispatch } from 'react-redux'
