+++ src/device/modalCleanup.ts
@@ -1,5 +1,5 @@
 /** Does some manual DOM manipulation to clean up after modals. */
-export const modalCleanup = () => {
+const modalCleanup = () => {
   const modalContainer = document.querySelector('.modal-container')
   if (modalContainer) {
     modalContainer.classList.remove('modal-container')
@@ -8,3 +8,5 @@ export const modalCleanup = () => {
     sibling.classList.remove('sibling-after')
   })
 }
+
+export default modalCleanup
