+++ src/util/index.ts
@@ -65,7 +65,6 @@ export { logWithTime } from './logWithTime'
 export { makeCompareByProp } from './makeCompareByProp'
 export { mergeThoughts } from './mergeThoughts'
 export { mergeUpdates } from './mergeUpdates'
-export { modalCleanup } from './modalCleanup'
 export { moveLexemeThought } from './moveLexemeThought'
 export { never } from './never'
 export { nextThought } from './nextThought'
